requirements.txt
requests>=2.23.0
geocoder>=1.38.1
pyyaml>=5.3.1
dotted-dict>=1.1.3
